# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Hotel Management API",
    "version": "1.0.0",
    "description": "API for hotel managment system"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "password"
                ],
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created"
          }
        }
      }
    },
    "/api/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "User login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "password"
                ],
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated successfully",
            "content": {
              "application/json": {
                "example": {
                  "accessToken": "ey...",
                  "refreshToken": "ey..."
                }
              }
            }
          }
        }
      }
    },
    "/api/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Refresh access token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "refreshToken"
                ],
                "properties": {
                  "refreshToken": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New access token returned"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Logout user (clear refresh token)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Logout successful"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/bookings": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "summary": "Get all bookings",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of all bookings"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Bookings"
        ],
        "summary": "Create a booking",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "guest_id",
                  "room_id",
                  "check_in_date",
                  "check_out_date"
                ],
                "properties": {
                  "guest_id": {
                    "type": "integer"
                  },
                  "room_id": {
                    "type": "integer"
                  },
                  "check_in_date": {
                    "type": "string",
                    "format": "date"
                  },
                  "check_out_date": {
                    "type": "string",
                    "format": "date"
                  }
                }
              },
              "example": {
                "guest_id": 1,
                "room_id": 2,
                "check_in_date": "2025-06-01",
                "check_out_date": "2025-06-05"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Booking created"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/bookings/{id}": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "summary": "Get booking by ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Booking found"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Booking not found"
          }
        }
      },
      "put": {
        "tags": [
          "Bookings"
        ],
        "summary": "Update a booking",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "guest_id": {
                    "type": "integer"
                  },
                  "room_id": {
                    "type": "integer"
                  },
                  "check_in_date": {
                    "type": "string",
                    "format": "date"
                  },
                  "check_out_date": {
                    "type": "string",
                    "format": "date"
                  }
                }
              },
              "example": {
                "guest_id": 1,
                "room_id": 2,
                "check_in_date": "2025-06-02",
                "check_out_date": "2025-06-06"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Booking updated"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Booking not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Bookings"
        ],
        "summary": "Delete a booking",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Booking deleted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Booking not found"
          }
        }
      }
    },
    "/api/bookings/available-rooms": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "summary": "Get available rooms by date",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "check_in_date",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "check_out_date",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Available rooms"
          },
          "400": {
            "description": "Invalid date format"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/bookings/revenue": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "summary": "Get total revenue by month",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "month",
            "required": true,
            "schema": {
              "type": "string",
              "example": "2025-06"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Revenue data"
          },
          "400": {
            "description": "Invalid month format"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/guests": {
      "get": {
        "tags": [
          "Guests"
        ],
        "summary": "Get all guests",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of all guests"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Guests"
        ],
        "summary": "Create new guest",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "full_name",
                  "email"
                ],
                "properties": {
                  "full_name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "full_name": "Vasyl Kisyl",
                "email": "vasylkisyl@i.ua"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Guest created"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/guests/{id}": {
      "get": {
        "tags": [
          "Guests"
        ],
        "summary": "Get guest by ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Guest found"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Guest not found"
          }
        }
      },
      "put": {
        "tags": [
          "Guests"
        ],
        "summary": "Update guest",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "full_name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "full_name": "Vasyl Kisyl",
                "email": "vasylkisyl@i.ua"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Guest updated"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Guest not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Guests"
        ],
        "summary": "Delete a guest",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Guest deleted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Guest not found"
          }
        }
      }
    },
    "/api/rooms": {
      "get": {
        "tags": [
          "Rooms"
        ],
        "summary": "Get all rooms",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of all rooms"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Rooms"
        ],
        "summary": "Create new room",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "number",
                  "type",
                  "price"
                ],
                "properties": {
                  "number": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  }
                }
              },
              "example": {
                "number": "101",
                "type": "Deluxe",
                "price": 120.5
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Room created"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/rooms/{id}": {
      "get": {
        "tags": [
          "Rooms"
        ],
        "summary": "Get room by ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Room found"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Room not found"
          }
        }
      },
      "put": {
        "tags": [
          "Rooms"
        ],
        "summary": "Update room",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "number": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  }
                }
              },
              "example": {
                "number": "101",
                "type": "Suite",
                "price": 150
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Room updated"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Room not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Rooms"
        ],
        "summary": "Delete a room",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Room deleted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Room not found"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Auth",
      "description": "Authentication routes"
    },
    {
      "name": "Bookings",
      "description": "Room bookings"
    },
    {
      "name": "Guests",
      "description": "Guest management"
    },
    {
      "name": "Rooms",
      "description": "Room management"
    }
  ]
}